+++ src-dbg/flow_storm/debugger/ui/main.clj
@@ -4,16 +4,29 @@
              :refer [event-handler h-box alert-dialog progress-indicator tab tab-pane border-pane]]
             [flow-storm.debugger.ui.flows.screen :as flows-screen]
             [flow-storm.debugger.ui.browser.screen :as browser-screen]
+            [flow-storm.debugger.target-commands :as target-commands]
+            [flow-storm.debugger.websocket :as websocket]
             [flow-storm.debugger.ui.state-vars
              :as ui-vars
-             :refer [main-pane stage scene obj-lookup store-obj dark-listener]]
+             :refer [obj-lookup store-obj]]
             [flow-storm.utils :refer [log log-error]]
-            [clojure.java.io :as io])
+            [clojure.java.io :as io]
+            [mount.core :as mount :refer [defstate]])
   (:import [com.jthemedetecor OsThemeDetector]
            [javafx.scene Scene]
            [javafx.stage Stage]
            [javafx.geometry Pos]
            [javafx.application Platform]))
+
+(declare start-ui)
+(declare stop-ui)
+(declare ui)
+
+(def flow-storm-core-ns 'flow-storm.core)
+
+(defstate ui
+  :start (start-ui (mount/args))
+  :stop  (stop-ui))
 
 (defn bottom-box []
   (let [box (doto (h-box [] "main-bottom-bar-box")
@@ -67,11 +80,6 @@
 
     tabs-p))
 
-(defn close-stage []
-  (when stage
-    (ui-utils/run-now
-     (.close stage))))
-
 (defn build-main-pane []
   (let [mp (border-pane {:center (main-tabs-pane)
                          :bottom (bottom-box)})]
@@ -83,10 +91,8 @@
         default-styles (if dark? "styles_dark.css" "styles.css")]
     (.clear stylesheets)
     (.add stylesheets (str (io/resource default-styles)))
-    (when styles (.add stylesheets (str (io/as-url (io/file styles)))))))
-
-(comment
-  (Platform/runLater #(update-styles scene true "styles_dark.css")))
+    (when styles (.add stylesheets (str (io/as-url (io/file styles)))))
+    nil))
 
 (defn reset-styles
   "Get current scene and a map config, containing (optionally)
@@ -94,10 +100,14 @@
   `:theme` — one of #{:light :dark :auto}, :auto by default
 
   Sets default style, overrides with extention `:styles` if provided.
-  Tries to remove a dark-listener. Sets new one if `:theme` is `:auto`."
+  Tries to remove a theme-listener. Sets new one if `:theme` is `:auto`.
+
+  If `:auto` is true, returns theme-listener, `nil` otherwise."
+
   [^Scene scene {:keys [styles theme] :or {theme :auto}}]
-  (let [detector (OsThemeDetector/getDetector)]
-    (when dark-listener (.removeListener detector dark-listener))
+  (let [detector (OsThemeDetector/getDetector)
+        theme-listener (:theme-listener ui)]
+    (when theme-listener (.removeListener detector theme-listener))
     (case theme
       :auto (let [detector (OsThemeDetector/getDetector)
                   dark? (.isDark detector)
@@ -105,17 +115,49 @@
                   listener (reify java.util.function.Consumer
                              (accept [_ dark?]
                                (Platform/runLater
-                                 #(update-styles scene dark? styles))))]
-              (alter-var-root #'dark-listener (constantly listener))
-              (.registerListener detector listener))
+                                #(update-styles scene dark? styles))))]
+              (.registerListener detector listener)
+              theme-listener)
       :light (update-styles scene false styles)
       :dark (update-styles scene true styles)
       (log-error "wrong `:theme`, should be one of: #{:light :dark :auto}"))))
 
-(defn reset-scene-main-pane []
-  (let [mp (build-main-pane)]
-    (alter-var-root #'main-pane (constantly mp))
-    (.setRoot scene mp)))
+(defn setup-commands-executor [{:keys [local?]}]
+  (if local?
+
+    ;; set up the local command-executor
+    (let [_ (require [flow-storm-core-ns])
+          run-command (resolve (symbol (str flow-storm-core-ns) "run-command"))]
+      (alter-var-root #'target-commands/run-command
+                      (constantly
+                       ;; local command executor (just call command functions)
+                       (fn run-command-fn [method args-map & [callback]]
+                         ;; run the function on a different thread so we don't block the ui
+                         ;; while running commands
+                         (.start (Thread. (fn []
+                                            (set-in-progress true)
+                                            (let [[_ cmd-res] (run-command nil method args-map)]
+                                              (set-in-progress false)
+                                              (if (= cmd-res :error)
+
+                                                (log-error "Error running command")
+
+                                                (let [[_ ret-val] cmd-res]
+                                                  (when callback
+                                                    (callback ret-val))))))))))))
+
+    ;; else, set up the remote command-executor
+    (alter-var-root #'target-commands/run-command
+                    (constantly
+                     ;; remote command executor (via websockets)
+                     (fn run-command-fn [method args-map & [callback]]
+                       (set-in-progress true)
+                       (websocket/async-command-request method
+                                                        args-map
+                                                        (fn [ret-val]
+                                                          (set-in-progress false)
+                                                          (when callback
+                                                            (callback ret-val)))))))))
 
 (defn start-ui [config]
   ;; Initialize the JavaFX toolkit
@@ -124,7 +166,11 @@
 
   (ui-utils/run-now
    (try
-     (let [scene (Scene. (build-main-pane) 1024 768)]
+     (let [scene (Scene. (build-main-pane) 1024 768)
+           stage (doto (Stage.)
+                   (.setTitle "Flowstorm debugger")
+                   (.setScene scene))
+           theme-listener (reset-styles scene config)]
 
        (doto scene
          (.setOnKeyPressed (event-handler
@@ -136,22 +182,26 @@
                                      (= key-name "G"))
                                 (do
                                   (log "Interrupting long running task")
-                                  (.interrupt @ui-vars/long-running-task-thread))
+                                  (ui-vars/interrupt-long-running-task-thread))
 
                                 ;; :else
                                 ;; (log (format "Unhandled keypress %s" key-name))
-                                )))))
+                                ))))
+         (.setRoot (build-main-pane)))
 
        (reset-styles scene config)
 
-       (alter-var-root #'scene (constantly scene))
-       (alter-var-root #'stage (constantly (doto (Stage.)
-                                             (.setTitle "Flowstorm debugger")
-                                             (.setScene scene)))))
+       (setup-commands-executor config)
+       (-> stage .show)
 
-     (reset-scene-main-pane)
-
-     (-> stage .show)
+       {:scene scene
+        :stage stage
+        :theme-listener theme-listener})
 
      (catch Exception e
        (log-error "UI Thread exception" e)))))
+
+(defn stop-ui []
+  (when-let [stage (:stage ui)]
+    (ui-utils/run-now
+     (.close stage))))