+++ src-dev/dev.cljs
@@ -80,23 +80,3 @@
   (fs-api/remote-connect {:on-connected (fn []
                                           (js/console.log "Connected!!")
                                           #rtrace (js/console.log (boo [2 "hello" 8])))}))
-
-;; (. bla clojure.core/addMethod :bla (clojure.core/fn [a] a))
-;; (cljs.core/-add-method bla :bla (cljs.core/fn [a] a))
-
-;; (comment
-;;   (def t (tt/->FnCallTrace 0 1 "foo" "foo-ns" 11 [42 41] 0))
-
-;;   (ser/serialize t)
-;;   (ser/deserialize (ser/serialize t))
-;;   (ser/deserialize (ser/serialize #"h.*"))
-
-;;   (fs-api/remote-connect)
-;;   (fs-api/close-remote-connection)
-
-;;   (fs-api/run {:ns "dev"}
-;;     (factorial 5))
-
-;;   "[\"~#flow_storm.trace_types.FnCallTrace\",[\"^ \",\"~:flow-id\",0,\"~:form-id\",1,\"~:fn-name\",\"foo\",\"~:fn-ns\",\"foo-ns\",\"~:thread-id\",11,\"~:args-vec\",[42,41],\"~:timestamp\",0]]"
-;;   "[\"~#flow_storm.trace_types.FnCallTrace\",[\"^ \",\"~:flow-id\",0,\"~:fn-name\",\"foo\",\"~:thread-id\",11,\"~:form-id\",1,\"~:fn-ns\",\"foo-ns\",\"~:args-vec\",[42,41],\"~:timestamp\",0]]"
-;;   )