+++ CHANGELOG.md
@@ -3,8 +3,16 @@
 ## master (unreleased)
 
 ## New Features
+
+	* New functions for shutting down the debugger and connections gracefully.
+	  When starting with `flow-storm.api/local-connect` or `flow-storm.api/remote-connect` you can shut it down with `flow-storm.api/stop` 
+	  When starting a standalone debugger with `flow-storm.debugger.main/start-debugger` you can shutdown with `flow-storm.debugger.main/stop-debugger`
+
+    * Add support for light and dark themes, in selected or automatic mode. Checkout the user guide for more info.
 	
 ### Changes
+
+	* The entire debugger was refactored to manage state with mount
 	
 ### Bugs fixed